refactor(vite): split out dev server configuration into plugin
