+++ packages/vite/src/client.ts
@@ -1,16 +1,13 @@
-import type { IncomingMessage, ServerResponse } from 'node:http'
 import { dirname, isAbsolute, join, relative, resolve } from 'pathe'
 import * as vite from 'vite'
 import vuePlugin from '@vitejs/plugin-vue'
 import viteJsxPlugin from '@vitejs/plugin-vue-jsx'
-import type { BuildOptions, ServerOptions } from 'vite'
+import type { BuildOptions } from 'vite'
 import { logger, useNitro } from '@nuxt/kit'
-import { getPort } from 'get-port-please'
 import { joinURL, withoutLeadingSlash } from 'ufo'
 import { defu } from 'defu'
 import { defineEnv } from 'unenv'
 import { resolveModulePath } from 'exsolve'
-import { createError, defineEventHandler, handleCors, setHeader } from 'h3'
 import type { Nuxt, ViteConfig } from '@nuxt/schema'
 
 import type { ViteBuildContext } from './vite'
@@ -22,6 +19,7 @@ import { ViteNodePlugin } from './plugins/vite-node'
 import { createViteLogger } from './utils/logger'
 import { StableEntryPlugin } from './plugins/stable-entry'
 import { AnalyzePlugin } from './plugins/analyze'
+import { DevServerPlugin } from './plugins/dev-server'
 
 export async function buildClient (nuxt: Nuxt, ctx: ViteBuildContext) {
   const nodeCompat = nuxt.options.experimental.clientNodeCompat
@@ -137,6 +135,7 @@ export async function buildClient (nuxt: Nuxt, ctx: ViteBuildContext) {
       // ensure changes in chunks do not invalidate whole build
       StableEntryPlugin(nuxt),
       AnalyzePlugin(nuxt),
+      DevServerPlugin(nuxt),
     ],
     appType: 'custom',
     server: {
@@ -148,18 +147,6 @@ export async function buildClient (nuxt: Nuxt, ctx: ViteBuildContext) {
   } satisfies vite.InlineConfig, nuxt.options.vite.$client || {}))
 
   clientConfig.customLogger = createViteLogger(clientConfig)
-
-  // In build mode we explicitly override any vite options that vite is relying on
-  // to detect whether to inject production or development code (such as HMR code)
-  if (!nuxt.options.dev) {
-    clientConfig.server!.hmr = false
-  }
-
-  // Inject an h3-based CORS handler in preference to vite's
-  const useViteCors = clientConfig.server?.cors !== undefined
-  if (!useViteCors) {
-    clientConfig.server!.cors = false
-  }
 
   // We want to respect users' own rollup output options
   const fileNames = withoutLeadingSlash(join(nuxt.options.app.buildAssetsDir, '[hash].js'))
@@ -180,25 +167,6 @@ export async function buildClient (nuxt: Nuxt, ctx: ViteBuildContext) {
     } satisfies NonNullable<BuildOptions['rollupOptions']>['output'],
   }) as any
 
-  if (clientConfig.server && clientConfig.server.hmr !== false) {
-    const serverDefaults: Omit<ServerOptions, 'hmr'> & { hmr: Exclude<ServerOptions['hmr'], boolean> } = {
-      hmr: {
-        protocol: nuxt.options.devServer.https ? 'wss' : undefined,
-      },
-    }
-    if (typeof clientConfig.server.hmr !== 'object' || !clientConfig.server.hmr.server) {
-      const hmrPortDefault = 24678 // Vite's default HMR port
-      serverDefaults.hmr!.port = await getPort({
-        port: hmrPortDefault,
-        ports: Array.from({ length: 20 }, (_, i) => hmrPortDefault + 1 + i),
-      })
-    }
-    if (nuxt.options.devServer.https) {
-      serverDefaults.https = nuxt.options.devServer.https === true ? {} : nuxt.options.devServer.https
-    }
-    clientConfig.server = defu(clientConfig.server, serverDefaults as ViteConfig['server'])
-  }
-
   await nuxt.callHook('vite:extendConfig', clientConfig, { isClient: true, isServer: false })
 
   clientConfig.plugins!.unshift(
@@ -208,83 +176,12 @@ export async function buildClient (nuxt: Nuxt, ctx: ViteBuildContext) {
 
   await nuxt.callHook('vite:configResolved', clientConfig, { isClient: true, isServer: false })
 
-  // Prioritize `optimizeDeps.exclude`. If same dep is in `include` and `exclude`, remove it from `include`
-  const exclude = new Set(clientConfig.optimizeDeps!.exclude)
-  clientConfig.optimizeDeps!.include = clientConfig.optimizeDeps!.include!
-    .filter(dep => !exclude.has(dep))
-
   if (nuxt.options.dev) {
     // Dev
     const viteServer = await vite.createServer(clientConfig)
     ctx.clientServer = viteServer
     nuxt.hook('close', () => viteServer.close())
     await nuxt.callHook('vite:serverCreated', viteServer, { isClient: true, isServer: false })
-    const transformHandler = viteServer.middlewares.stack.findIndex(m => m.handle instanceof Function && m.handle.name === 'viteTransformMiddleware')
-    viteServer.middlewares.stack.splice(transformHandler, 0, {
-      route: '',
-      handle: (req: IncomingMessage & { _skip_transform?: boolean }, res: ServerResponse, next: (err?: any) => void) => {
-        // 'Skip' the transform middleware
-        if (req._skip_transform) { req.url = joinURL('/__skip_vite', req.url!.replace(/\?.*/, '')) }
-        next()
-      },
-    })
-
-    const staticBases: string[] = []
-    for (const folder of useNitro().options.publicAssets) {
-      if (folder.baseURL && folder.baseURL !== '/' && folder.baseURL.startsWith(nuxt.options.app.buildAssetsDir)) {
-        staticBases.push(folder.baseURL.replace(/\/?$/, '/'))
-      }
-    }
-
-    const devHandlerRegexes: RegExp[] = []
-    for (const handler of nuxt.options.devServerHandlers) {
-      if (handler.route && handler.route !== '/' && handler.route.startsWith(nuxt.options.app.buildAssetsDir)) {
-        devHandlerRegexes.push(new RegExp(
-          `^${handler.route
-            .replace(/[.+?^${}()|[\]\\]/g, '\\$&') // escape regex syntax characters
-            .replace(/:[^/]+/g, '[^/]+') // dynamic segments (:param)
-            .replace(/\*\*/g, '.*') // double wildcard (**) to match any path
-            .replace(/\*/g, '[^/]*')}$`, // single wildcard (*) to match any segment
-        ))
-      }
-    }
-
-    const viteMiddleware = defineEventHandler(async (event) => {
-      const viteRoutes: string[] = []
-      for (const viteRoute of viteServer.middlewares.stack) {
-        const m = viteRoute.route
-        if (m.length > 1) {
-          viteRoutes.push(m)
-        }
-      }
-      if (!event.path.startsWith(clientConfig.base!) && !viteRoutes.some(route => event.path.startsWith(route))) {
-        // @ts-expect-error _skip_transform is a private property
-        event.node.req._skip_transform = true
-      } else if (!useViteCors) {
-        const isPreflight = handleCors(event, nuxt.options.devServer.cors)
-        if (isPreflight) {
-          return null
-        }
-        setHeader(event, 'Vary', 'Origin')
-      }
-
-      // Workaround: vite devmiddleware modifies req.url
-      const _originalPath = event.node.req.url
-      await new Promise((resolve, reject) => {
-        viteServer.middlewares.handle(event.node.req, event.node.res, (err: Error) => {
-          event.node.req.url = _originalPath
-          return err ? reject(err) : resolve(null)
-        })
-      })
-
-      // if vite has not handled the request, we want to send a 404 for paths which are not in any static base or dev server handlers
-      if (!event.handled && event.path.startsWith(nuxt.options.app.buildAssetsDir) && !staticBases.some(baseURL => event.path.startsWith(baseURL)) && !devHandlerRegexes.some(regex => regex.test(event.path))) {
-        throw createError({
-          statusCode: 404,
-        })
-      }
-    })
-    await nuxt.callHook('server:devHandler', viteMiddleware)
   } else {
     // Build
     logger.info('Building client...')
